FIXED AN ISSUE WHERE THE APP WOULD CRASH AFTER EDITING OR DELETING NEWLY INSTANCED ENTRIES.

+++ app/src/main/java/com/cc221012_cc221016/stash/data/EntriesDao.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface EntriesDao {
     @Insert
-    suspend fun insertEntry(entry: Entries) : Long
+    suspend fun insertEntry(entry: Entries) :Long
 
     @Update
     suspend fun updateEntry(entry: Entries)
@@ -22,5 +22,5 @@ interface EntriesDao {
     fun getEntries(): Flow<List<Entries>>
 
     @Query("SELECT * FROM entries WHERE entryID = :id")
-    fun getEntry(id: Int): Flow<Entries>
+    fun getEntry(id: Long): Flow<Entries>
 }
